requirements.txt
azure-functions>=1.18.0
pyodbc>=5.1.0
pandas>=2.2.2
numpy>=1.26.4
openai>=1.40.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
